add: feature to specify random seed argument
I've put it in default configuration files, now there's an option `random_seed`. If it's `None`, random seed is different across restarts. If it's a constant integer, calls of random.random() function will produce the same sequence of values. By now it's only used for generation of captcha codes to make it possible to write functional tests for the correcponding endpoint.